Support TXT files as potential gallery targets
Right-clicking on a txt file in file explorer will
show an option to open it as a gallery.

Such TXT files usually contain images outside of the working directory,
therefore VSCode is given permission to access all files on the local
system. <- This could be reduced to parent folders of all images.